+++ src/nrepl/repl-start.ts
@@ -10,7 +10,7 @@ import { getConfig } from '../config';
 import * as replSession from './repl-session';
 import * as cljsLib from '../../out/cljs-lib/cljs-lib';
 import { ReplConnectSequence } from './connectSequence';
-import * as clojureLsp from '../lsp/main';
+import * as clojureLsp from '../lsp';
 import * as calvaConfig from '../config';
 
 const TEMPLATES_SUB_DIR = 'bundled';
@@ -166,9 +166,6 @@ export async function startStandaloneRepl(
       console.error(`Error downloading drams: ${err.message}`);
       void vscode.window.showWarningMessage(`Error downloading files: ${err.message}`);
     });
-  }
-  if (clojureLsp.lspStatus === 'stopped' && calvaConfig.getConfig().enableClojureLspOnStart) {
-    void clojureLsp.startClientCommand();
   }
 
   const [mainDoc, mainEditor] = await openStoredDoc(storageUri, tempDirUri, config.files[0]);