+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- [LSP support for multi-project and multi-workspace](https://github.com/BetterThanTomorrow/calva/pull/2020)
+- Fix: [Clojure-lsp caching in multi-project workspace](https://github.com/BetterThanTomorrow/calva/issues/934)
+- Fix: [With multi-root workspaces the clojure-lsp project root is always the first folder](https://github.com/BetterThanTomorrow/calva/issues/1706)
 - Fix: [Getting Started REPLs download config from dram dev branch](https://github.com/BetterThanTomorrow/calva/issues/1977)
 
 ## [2.0.326] - 2023-01-24